добавлена страница логина, структура для pages
