+++ src/components/App/App.jsx
@@ -1,31 +1,30 @@
 import React, { useEffect } from 'react';
 import styles from './App.module.css';
 import AppHeader from '../AppHeader/AppHeader';
-import BurgerIngredients from '../BurgerIngredients/BurgerIngredients';
-import BurgerConstructor from '../BurgerConstructor/BurgerConstructor';
-import { useDispatch, useSelector } from 'react-redux';
-import { getItemsData } from '../../services/actions/ingresients-data';
-import { DndProvider } from "react-dnd";
-import { HTML5Backend } from "react-dnd-html5-backend";
+import MainPage from '../../pages/main/main';
+import Login from '../../pages/login/login';
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 
 function App() {
-  const itemsLoaded = useSelector((store) => store.ingredients.itemsLoaded);
-  const dispatch = useDispatch();
-  useEffect(() => { dispatch(getItemsData()) }, [dispatch])
 
   return (
     <div className={styles.App}>
       <AppHeader />
-      {itemsLoaded === false
-        ? <div>Загрузка...</div>
-        : <main className={styles.AppMain}>
-          <DndProvider backend={HTML5Backend}>
-            <BurgerIngredients />
-            <BurgerConstructor />
-          </DndProvider>
-        </main>
-      }
+      <Router>
+        <Switch>
+          <Route path="/" exact={true}>
+            <MainPage />
+          </Route>
+          <Route path="/login" exact={true}>
+            <Login />
+          </Route>
+          <Route path="/register" exact={true}>
+
+          </Route>
+        </Switch>
+      </Router>
     </div>
+
   );
 }
