+++ src/pages/login/login.module.css
@@ -1,0 +1,37 @@
+.login {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 0;
+  width: 480px;
+  height: 430px;
+  margin: 180px auto 0;
+  border: 1px dashed lightcyan;
+}
+
+.title {
+  margin: 0;
+}
+
+.form {
+  display: grid;
+  gap: 24px;
+  padding: 24px 0 80px 0;
+  justify-items: center;
+}
+
+.newUser {
+  margin: 0;
+  padding: 0 0 16px 0;
+  color: #8585AD;
+}
+
+.resetPassword {
+  margin: 0;
+  color: #8585AD;
+}
+
+.link {
+  text-decoration: none;
+  color: #4C4CFF;
+}
