+++ src/pages/main/main.module.css
@@ -1,0 +1,7 @@
+.AppMain {
+  display: flex;
+  flex-direction: row;
+  gap: 40px;
+  justify-content: center;
+  padding: 40px 0 0;
+}
